Redirect admin user to admin context (#530)
#### Current state

When accessing the `admin_users/sign_in` page, the user will be redirected to the root page "/", instead of the "admin" context.

#### What has changed?

Now the user will be redirected to the admin context, but old behavior sill works. If another admin page has been accessed, the user will be redirected to that instead.